Make the T_gaia object ctor that takes an id private
This ctor that takes a gaia id should only be callable from the private
get_object helper function that knows how to create the object with an
id and its associated flatbuffer.  Direct user access to this constructor
causes an object to be built that can't be read from since there is
no underlying flatbuffer.