[GAIA-28]  Event manager implementation (#23)
* added list_subscribed_rules method, added unsubscribe_rules method to clear all rules, cleaned up tests to verify rule context values as well as added tests for added APIs

* Switch error handling from codes to exceptions

Added dedicated exceptions for duplicate_rule, invalid_event_type and
mode_not_supported.  Log event methods also return true or false to
indicate whether at least one rule was fired in response to the event.

* Define that a rules application must provide a intialize_rules() function

The event_manager singleton will call initialize_rules() on construction.
This means that the rules application (and tests) must provide an implementation.

If a rule calls log_event with the same event that caused that
rule to be fired then we want to suppress further invocations
of that rule for Q1.

* Prevent reentrant calls to the same event

Add a guard to prevent reentrant call to the log_event
method while processing the same log_event.  This prevents
a call cycle when a rule directly calls log_event or indirectly
calls it by doing an action that causes a log_event
to be made.