Add libraries for system storage engine, and direct access
Pull definitions out of headers and link to them as libraries.  Clients
have the choice of linking to gaia_system which will pull in the other
link targets or manually linking to gaia_se and gaia_direct for
storage engine and direct access definitions.  The system library
has one initialization routine for now that initializes the
event manager and storage engine.