Removed gaia_type parameter from log_*_event APIs
Added support in the gaia_object_t to pick up the gaia type from the
template specialization.  This enables using a static gaia type MyObject::s_gaia_type
when you don't have an instance (the rule subscriber APIs).  Also added
a runtime virtual method gaia_type() so that the event manager can
pick up the time when handed an instance of a gaia_base_t.  With this info
the user no longer needs to pass the explicit gaia_type to the event methods.