Address code review feedback
Also rename event_tx_t to auto_tx_t.  This class is used to
start a transaction if one is not already active. If the function
fails (by calling a routine that throws an exception, for example) then
the destructor of this class will rollback the transaction.  The
client must explicitly commit the transaction now.

Added tests around this class and a test case to event_manager to
verify that a transaction is created if an event is logged without
a transaction context.