Postgres FDW implementation for COW-SE (#78)
This commit implements a native C++ Foreign Data Wrapper (FDW) for Postgres, exposing the Gaia storage engine to the Postgres query processor.

The current implementation hardcodes the schema for the Q1 demos (the airport demo tables and the Gaia event log table), both in the query and modify methods and in the CREATE FOREIGN SCHEMA implementation. There is no way to use the current implementation with any other types, unless the implementation is modified to support them. An aspirational goal for Q2 is to infer the relational schema for all tables from a FlatBuffers binary schema, and use the FlatBuffers reflection API to implement all record accesses and modifications.

Testing is currently just 4 SELECT queries from the Q1 demo, but needs to be expanded to cover INSERT/UPDATE/DELETE queries, and possibly unit tests. The file `/GaiaPlatform/production/sql/tests/setup.sql` can be used as an example of how to use the foreign data wrapper in SQL queries.