Camera demo packaging (#83)
* Make cameraDemo relocatable

Some of the paths, such as location of the OpenCV cnn data files had
paths calculated at compile time. We expect them to be located in the
same folder as our binary. We now calculate that folder at runtime.

Additionally, the staging area for captured images is changed from CWD
to /tmp, since CWD will be a read-only filesystem in some cases (such as
after the camera demo is packaged in an AppImage).

* Fix OpenCV in cameraDemo CmakeLists.txt

* Add cameraDemo packaging

Running run.sh uses docker to build an AppImage of camera demo.

* Add camera demo packaging README.md