Rules invocation and mock trigger function (#93)
Remove LogEvent APIs in favor of the mock trigger function.

Note that the trigger function will be invoked after commit by the storage
engine in the real implementation.  The mock is only added to test the 
rule scheduler side of the API to map trigger events to subscribed rules.

* Remove LogEvent APIs as well as extra rule subscription APIs.

Q2 events are sourced by storage engine triggers so there is no need for
an explicit LogEvent API.  Also changed the way rules are bound to events
to enable binding to fied ids that represent fields referenced in a
declarative rule.

* Added immediate execution mode for rule thread pool

If the thread pool is created with 0 threads then all the rules
are execute on the same calling thread.  This enables deterministic
testing for the time being.  Also brought more tests back online
with the new rules APIs.

* Continue update of event manager tests

* Last round of event_manager test updates; attach Jira item ids to TODO comments

* Address code review feedback

* Simplify thread init function

* Reorgnaize trigger function code in response to review feedback

* Change condition variable name to be more descriptive