Enable demos to build off Q2 rule engine APIs
Modify the gaiaflatc compiler to emit mock commit_trigger
calls and simulate the post-commit firing of rules.  Also
modify the incubator demo to work with this new Q2 model.

Although binding rules to transaction events has not been
specified for Q2, enable a transaction rollback will abandon
any queued events so that rules are not fired.