Introduce a simplified API and address gaia_cache memory issues
The simplified API allows calls to insert_row() and update_row() to
be omitted when used with a helper "auto transaction" object.  The
semantics of the transaction object require an explicit commit but
then immmediately start a new transaction.  Prior to commit, the
transaction object will give the opportunity to each EDC class to
do "pre-commit" work (such as the insert_row or update_row operations).
No new functionality (besides the ability to omit explict comands) is
introduced by the simplified API.

The memory issues addressed include the following changes.  A user
must now use a Type::create() method to create a new object instead of
the using "new".  The return type of create() and all the APIs that
retrieve obejcts (get_first, get_next, get_row_by_id) are now a
shared_ptr<Type>.

EDC now has lifetime control of the objects in the gaia_cache.  This
means the cache now knows whether objects in the cache have an
external reference (so therefore must be refreshed across transaction
boundaries) or whether it is safe to "garbage collect" them because the
only reference is held by the cache itself.

The TX specific gaia cache has been removed as we still needed to refresh
objects that may have not been involved in the transaction but still have
outstanding references.

An temporary object cache has been added to track new objects that do not
have a record id.  Unlike the gaia cache, the object cache only holds weak
references so it does not keep the object alive.  When the object is inserted
(attached to the database) it is moved to the gaia cache where it is tracked.
If the object is deleted before the insert_row then the system forgets about
it.