Require separate writer objects for inserting and updating rows
This change simplifies the read path for EDC objects over flatbuffer row data.
When a row is inserted, the user must now create a "writer" object.  Writer
objects are passed into static methods for inserts and updates.  This means
that the gaia object no longer has to handle state where a new object changes
to a database-backed object after insert.  This also means that no setters
or original value methods are surfaced on the generated object.  The writer object
is the "objT" structure generated by the flatc compiler.  A "get_writer" method
has been added to the gaia_object_t class.  This returns the objT structure
corresponding to the flatbuffer and copies the exist values into the copy.

In addition, overloads of insert_row and delete_row exist that enable insertion
and deletion without needing to create any object first.