Init catalog change (#96)
This pull request sets up the catalog project. It includes the catalog library and a simpler parser for catalog DDLs.

The library only includes APIs for DDLs. We will expand the library to build out more APIs in future changes.

The lexer and parser syntax are defined in .ll and .yy files, and will be compiled by flex and bison to generate the scanner and parser code. The main.cpp in the catalog project implements a simple tool for demonstration and manual testing purpose.

The design doc for the catalog can be found at:
https://docs.google.com/document/d/1KEUen8Til-_Tcu2Q6M7HzxrV4D6re8vEXKwC4gtTeS0/edit