Enable caching of the writer object on the type instance object
This change makes update_row an instance method instead of a static
method.  The writer returned from an instance of a gaia object is
now owned by the gaia object via a reference count on a shared ptr
object.  This allows multiple calls to instance->writer()->field followed
by a call to instance->update_row().  The prior API required that
the user have a writer object in hand which made it less intuitive.