Catalog foreign key support (#126)
This change introduces the foreign key support in Gaia catalog. Foreign keys are specified using REFERENCES keyword when creating tables. A new API list_references() is introduced to list references for a given table. The references for a table is ordered separately with other data fields. They will also not show up in the generated FlatBuffer schema (either binary or plaintext).

Forward reference declaration is not supported in this change. The work is tracked in the https://gaiaplatform.atlassian.net/browse/GAIAPLAT-258