Change the event manager tests to run synchronously.
Exposed a private initialize_rules_engine method under
gaia::rules::test namespace to enable creating the rules engine with
a passed in number of threads for the rule scheduler thread pool.  In
addition, expose a private commit_trigger method for for simulating
the trigger call synchronously.  This allows us to test the logic
of the event manager without having any background threads.