Enable the auto_transaction_t class to be safe for nesting.
Code should be able to instantiate an instance of the auto_transaction_t class without having
to worry about it throwing an exception because a transaction has already been started.  This change
makes any operations on a nested instance of auto_transactino-t a no-op.  We enable the scenario
when a rule wants to commit a transaction that the rules engine started by passing a reference
to the transaction in the rule context now.