Move field_position_t type to gaia_common header. (#175)
The field_position type and list type are used by both the database and
the rules engine.  This change removes the rule specific field_list_t type
and replaces it with the field_position_t type.  Because this type is now
used in the public rules engine api, it has been moved under inc/public..