Enable default construction of EDC objects. (#182)
This allows code to use EDC objects as instances of other classes
and structs and then set the reference later.  This is merely a
convenience method to allow the classes to be used more easily within
an application.  An object created in this way will have an underlying
gaia_ptr of nullptr so every non-static method will throw with an
invalid node id exception.