[gdev] support docker image registry (#187)
This allows us to upload images to the docker image registry via the
`gdev push` command. Additionally, `gdev build` will automatically query
the registry for cached build stages. The address of the registry may be
specified via the `--registry` flag. It currently defaults to
"192.168.0.250:5000", the registry on our build machine. Using the
registry requires us to use the VPN. If we're not connected to the VPN,
the CLI still works, but does not benefit from the image cache.

Using the registry currently requires that you enable unsafe registries
in the docker daemon. See the README.md for details.

There are additional, quality-of-life changes in here that I'm not
splitting out for the moment.
* Replaced `--sticky-mixins` and `--sticky-mounts` with `--mixins` and
`--mounts` flags respectively. The sticky behaviour turned out to be
more of a foot-gun than a help.
* Better image tag management. We now do a better job of pruning stale
images.