[GAIAPLAT-319] Support "CREATE IF NOT EXISTS" (#213)
Add support for "CREATE TABLE IF NOT EXISTS" and "CREATE DATABASE IF NOT EXISTS". Many databases already support this syntax. Adding this to the DDL helps both unit tests and future development. Unit tests right now uses a gaiac -e option to bypass the exceptions for creating tables or databases that already exists. I think the DDL is a better place to support the feature.