Log a tx when inserting or erasing references (#218)
Originally the change was only intended to log a tx for reference update. It has since expanded the scope to include changes for the following items:

[GAIAPLAT-354] Reference updates need to log transactions
[GAIAPLAT-355] Reference updates could break snapshot isolation semantics
[GAIAPLAT-356] [EDC] gaia_iterator template slot types should be size_t
[GAIAPLAT-357] [EDC] rename primary and foreign to parent and child in gaia_iterator

A unit test is added to test_recovery (converted to googletest but not added to ctest) to cover GAIAPLAT-354.
A unit test in test_references is updated to conform with the new reference update transaction semantics.