Use cmake ExternalProject to install headers only deps (#239)
I did some investigation and found a way to manage lightweight C++ dependencies via cmake ExternalProject and INTERFACE libraries. This change is compatible with existing gdev based build. If a dependency is found installed either via gdev or manually on the OS, the dependency setup step will be skipped. Otherwise, cmake will automatically download and install the dependency in the build directory as part of the local build.

The main motivation for this change is to make it easier to build locally without gdev as the lightweight C++ dependencies keep growing. The scope is limited to lightweight C++ dependencies (that also uses cmake to manage their builds). This approach shall never apply to bigger dependencies like PostgreSQL, RocksDB, or LLVM.

The following dependencies are covered in this change:

- backward
- cpptoml (spdlog_setup dependency)
- fmt (spdlog dependency)
- spdlog
- spdlog_setup
- tabulate

For adding C++ dependencies in the future, CMake is not required. gdev is still the choice of build machine and will cover all dependencies.