Rules stats counters (#276)
* Add statistics gathering and logging to the rules engine

The rules engine now gathers two types of statistics: scheduler and rules.
The scheduler statistics are always gathered and aggregate the counters for
all rules that are run.  Individual rule statistics must be enabled by the
user and are gathered for each rule that is executed.  The user can control how
often the statistics are written to a separate gaia_stats log file by setting
the log_interval (in seconds) on the event_manager_settings class. If the
interval is to 0 seconds then no log file entries are written.

* Simplify stats logging and use a tabular format for output