Add light_catalog to storage engine (#277)
This change adds a lightweight API to read catalog in storage engine without involving direct access APIs.

There are minor refactoring/reformat here and there. The main things are implemented in
- light_catalog.hpp
- light_catalog.cpp

The idea is that catalog API in storage engine only knows limited catalog knowledge and can be used to read the catalog records in shared memory storage. Right now, the implementation uses the following information of the catalog.
- gaia_table payload schema in order to load fbs schema for tables.
- gaia_field payload schema and references slot knowledge in order to iterator fields.

The cache that holds type to table id mapping is moved from ddl_executor to storage engine.

payload_diff on longer needs to rely on catalog direct access APIs, and we are rid of that circular dependency.

To unblock index change, it should be relatively straightforward to make light_catalog (comparing with direct access API) work on the server side.