Add tests to cover rule_stats_manager, scheduler_stats, and rule_stats (#298)
* Add tests to cover rule_stats_manager, scheduler_stats, and rule stats

Introduce a debug logger to facilitate log output testing.  Refactor
the underlying stats classes to use dependency injection for the logger
and move several private members to protected to aid in testing.

* Enable overriding the global rule_stats logger for testing

This removes the requirement to pass in a logger from the event manager all the way down
rule stats class.