Add system configuration initialization
Gaia system library will now parse a gaia_conf configuration file.
It then will hand off the root of the configuration file to each
sub-component so that it can handle its own section.  If no configuration
file is passed in to the system initialize call, then an empty root
is passed.