Fix bug in EDC containers generation code (#327)
In EDC the parent side of a relationship exposes the children via reference_chain_container_t. reference_chain_container_t is returned by reference, but EDC classes are often instantiated on the stack. If the object goes out of scope the reference to reference_chain_container_t will be invalid.

EDIT: Note that this could happen even if the object was heap-allocated, but being the object stack-allocated it makes the section of the problem a little harder.

This PR fixes the problem returning copies of reference_chain_container_t instead of references.