Remove reference typed fields in catalog gaia_field (#334)
This change removes referenced typed field in catalog gaia_field table. The references for tables are defined in gaia_relationship table @simone-gaia has created.

Also, updated EDC generation slightly to define type aliases for container lists. So it will be easier to use container list types in code that are not auto generated.