Move internal headers into public for the SDK
This change moves internal headers that are referenced by public headers
into public.  Biggest changes is that the gaia_ptr class becomes public.
The gaia_se_object remains internal by moving some implementation from the
header into the cpp file.

Build the SDK even in non-GaiaRelease builds to ensure an app can
link to the shared library and is not pulling in headers that are not internal.