Move RI from EDC to SE (#352)
The main purpose of this PR is to move the Referential Integrity (RI) logic from EDC to the Storage Engine (SE), leveraging the new APIs provided in gaia_ptr. 

## Schema loading

One of the outcomes of moving RI to SE is that the Catalog MUST contain the tables that you use to manipulate data. For instance, if a test uses the `employee_t` class, the table `employee` must exist in the Catalog, otherwise, the test will likely fail because the metadata will not contain the information about the `employee` class. 

To facilitate this I have added:
- schema_repository_t: A class that helps to manage DDL files. Given a file name, it can search for it in a list of predefined paths.
- db_catalog_test_base_t: A base class for tests that handle the lookup and loading of DDL files (using schema_repository), and clean up the DB status after every test. 

This lead to:
- Copy all the scheams from schemas/test to the CMake build directory to make them discoverable by schema_repository_t
- Refactoring tests to use `db_catalog_test_base_t`.
- Marking some header-only methods `inline` to avoid "multiple definitions" link error.

## Type metadata

- The type_metadata_t interface has been simplified. 
- Fix bug in type_registry_t where not all the tables were loaded in the metadata
- Now the registry loads all the Catalog base tables (table, field, relationship, etc..)
- Add log statements 

## Miscellanea
- Made the ddl_executor_t::init() public so that the Catalog can correctly be initialized when reloading the Schemas and renamed it to reset().
- Remove `update_child_references` API from `gaia_ptr`
- Add log statements in ddl_executor
- Fix memory leak in the Lexer discovered with ASAN