Omit logging the internal "log_event" rule for rule statistics.
This checkin also includes the following changes minor changes for file
cleanup from a past pull request and a fix in the translation engine to
ensure rule ordinal ordering starts at 1.  The ordinal of the rule
corresponds to the order the rule is encountered in the ruleset and
is used as the "rule name" in the rule statistics.