Fix Linking error for rule_context_t::last_operation (#397)
The Linker did not export rule_context_t::last_operation because is not used anywhere outside its own module. The linker does not consider single functions but entire object files for export (unless specified otherwise). The fix is to move this function within a cpp file that is used somewhere: event_manager.cpp.