Provide a weak reference for initialize_rules() function.
The rules engine requires a user-supplied initialize_rules() function to
subscribe rules.  This function is called at rules engine initialization and
provided by the translation engine.  If the user wants to write an application
with no rules then provide a weak reference for this function.  If the user
supplies a definition of initialize_rules() then the strong reference will
replace the weak one.

Also added tests to verify that the link behavior is working as expected both
for static and dynamic linking.