Support vector of basic types (#536)
Add support of vector of basic types. Vectors of strings are not supported.

A few unrelated test files are modified due to the conflict of direct_access
vector and std vector.

Users can declare a field in the table as a vector of basic type in the
following syntax: CREATE TABLE rock_band (member_ids uint64[]);

Notes on writing and reading of the vector payload in EDC.

Creating or updating the vector type field is not copy free. To create or update
a row containing vector fields, users need to pass in (a pointer to) std::vector
of basic types. Per flatbuffers::CreateVector implementation, the data will be
copied element by element to the fbs payload when constructing the se object.

Reading the vector field is copy free. The returned type of the vector field is
a pointer to flatbuffers::Vector objects. Created GAIAPLAT-639 to track the work
to hide fbs vector.

Querying and modifying array types are supported in FDW. Vectors of int8 and
uint8 are mapped to smallint[] type in PostgreSQL instead of bytea. Created
GAIAPLAT-678 to track if we want to use bytea for int8 and uint8 vectors.