[GAIAPLAT-883] Embed spdlog/spdlog_setup as gaia_spdlog/gaia_spdlog_setup (#688)
The motivation for this change is to avoid Undefined Behavior when multiple versions of spdlog are used within the same application.

Example: ROS2 embeds its own version of spdlog which is likely to be different from the spdlog version used by Gaia. This causes problems when linking spdlog both as a static and dynamic library.

Changes:

- Embed spdlog within the Gaia sources and rename it gaia_spdlog (the idea was initially proposed by Chuan). This way there will not be any conflict between the Gaia spdlog version and our customers' version.
- Embed spdlog_setup sources and change them to refer to gaia_spdlog instead of spdlog. I renamed spdlog_setup into gaia_spdlog_setup.
- Changed the spdlog embedded version of fmt into gaia_fmt.

Missing:

Add credits to spdlog/spdlog_setup in the Gaia SDK (there are probably other OS libraries that we use that require to do so, this is a good time to do it) => Created follow-up task:https://gaiaplatform.atlassian.net/browse/GAIAPLAT-990

I contacted the authors of FMT and SPDLOG and they are both ok with the renaming as long as we give credit (the MIT license allow to do so).