[GAIAPLAT-672] Introduce syntax support for Connect()/Disconnect() in clang (#756)
Introduce the syntax support for Connect/Disconnect in the translation engine. This change requires the introduction of the concept of link, which is the name of a link going from one table to another. Links are a new source of "ambiguity" because now a name could be a table, a field, or a link.

As Gregory pointed out some of this ambiguity can be ruled out based on context, (eg, farmer->crop farmer must be a table in this context), but we are not there yet. There is a JIRA to track this.

Miscellanea:
- Run the entire translation process within a single transaction.
- Improved messaging when the server is not running.

Missing:
- Connect/Disconnect Parameters checking. (will be addressed in different PR)
- Actual translation logic. (will be addressed in different PR)