[GAIAPLAT-1167] Make connect/disconnect insert/remove arguments as const ref (#837)
This code snippet:

```
 farmer.connect(raised.insert(birthdate: "2 Aug 1990"));
```

Cased this error: `candidate function not viable: expects an l-value for 1st argument`. I thought this was a problem with the translation engine itself until that this is an intended CPP behavior: a non-const reference cannot bind to a temporary object.

We have the same problem in EDC where the following code does not compile:

```
emp.addresses().insert(address_t::get(address_t::insert_row("", "", "", "", "", "", true)));
```

This PR simply put the const modifier to connect/disconnect insert/remove arguments.

I also removed the AST checks for connect/disconnect because they do not provide much value and are really time-consuming to maintain.