Default DB Fixes (#907)
Fix `gaiat` and `process_schema` for the default database.

For default db, `gaiat` should not generate the `::[db_name]` as part of direct access class namespaces.

CMake function does not accept an empty string as argument. Even if `DATABASE_NAME ""` is used as named argument, it will still be treated as undefined per CMake usage. We changed `process_schema` to use the default database when `DATABASE_NAME` is undefined. As a result, most places where it is used are updated to pass in the database name explicitly.