Fix an ubsan error (#1143)
Fix for https://gaiaplatform.atlassian.net/browse/GAIAPLAT-1815

I think this is not a real issue because we won't have null vectors in fbs payload. On missing input or nullptr input, we will insert an empty vector right now. Nevertheless, this should fix the ubsan error and probably help with the null support in the future.