Revert "Define & implement the DDL for new catalog schema (#1141)" (#1166)
This reverts commit fd846aa.

The blocking technical reason to revert the change is because the type_metadata is not set for all new relationships that connect to the core catalog tables. As a result, all core catalog records are allocated in wrong size, i.e. they are allocated without all the added relationship reference slots. Any attempts to make connections to them will result in segfault.

Some high level thinking on the catalog schema design: we should consider removing the relationships to the core tables, and use values in other tables for navigation, e.g. instead of creating a relationship between rule_table and gaia_table, we can add a table_id field in rule_table. It is less convenient to navigate but it makes the core and bootstrap simpler. Based on previous experience with database catalog internals, this is worth the effort. gaia_index and gaia_realtionship are already doing this for gaia_field.