Fix broken includes in Cmake translate_ruleset_internal() (#1197)
This PR fixes the failure in Debug built introduced by https://github.com/gaia-platform/GaiaPlatform/pull/1165. The fix is to remove the libc++ headers that apparently never worked because of a trailing `-I`.

eg:
```
-I GaiaPlatform/production/cmake-build-gaiarelease-debug-make/gaia_generated/barn_storage  \
-I  \  <==================================== This was preventing the libc++ include to be included
-I /usr/lib/llvm-13/include/c++/v1/  \
-I /usr/include/c++/10  \
-I /usr/include/x86_64-linux-gnu/c++/10  \
-I /usr/include/c++/10/backward  \
```

The `-I` problem was fixed as part of a later command and hence the presence of the libc++ headers started to make the build fail.