[GAIAPLAT-1391] Allow deleting connected nodes (#1215)
Previously, if a node is connected to some other node(s), we will not allow users to delete it. This change enables user to delete the node even if it is still connected. The only exception is a connected parent node in explicit one to many relationships, in which case we will throw the same exception as before. A force option is added to the delete_row API to allow users to delete the parent node in this case. Note: force deletion is different from cascade deletion. After force deletion, all child nodes will be disconnected to the deleted parent node but not deleted.