[GAIAPLAT-1949] Create optional class as replacement of c++17 std::optional (#1262)
Implement a custom `optional` class to be used in DA classes to denote optional values. We can't use `std::optional` because that's a C++17 feature. 

This `gaia::common::optional_t` exposes a subset of the `std::otpional` API. This will allow in future to use `std::optional` directly (eg. like flatbuffers do).